Added MenuController Script.
Added a script to StartMenu to control it using public toggle functions.
Set the canvas to not show on start.
Added button function to start button to toggle the StartMenu canvas on and off.
Added Hide function to buttons inside StartMenu so that after they click and open another window they close the StartMenu.